+++ tests/testthat/helper-recipes.R
@@ -11,3 +11,10 @@ skip_if_recipes_post_0.1.16 <- function() {
     message = "Skipping on new recipes where no longer applicable (post 0.1.16)"
   )
 }
+
+skip_if_recipes_pre_0.2.1 <- function() {
+  skip_if(
+    condition = packageVersion("recipes") < "0.2.0.9001",
+    message = "Skipping on old recipes (pre 0.2.1)"
+  )
+}